Moved VersionDateTime (removed duplicated code)
