+++ whois-commons/src/main/java/net/ripe/db/whois/common/dao/VersionLookupResult.java
@@ -2,7 +2,6 @@ package net.ripe.db.whois.common.dao;
 
 import net.ripe.db.whois.common.domain.serials.Operation;
 import net.ripe.db.whois.common.rpsl.ObjectType;
-import net.ripe.db.whois.query.VersionDateTime;
 
 import javax.annotation.concurrent.Immutable;
 import java.util.Collections;
