+++ whois-query/src/main/java/net/ripe/db/whois/query/domain/DeletedVersionResponseObject.java
@@ -2,7 +2,7 @@ package net.ripe.db.whois.query.domain;
 
 import net.ripe.db.whois.common.domain.ResponseObject;
 import net.ripe.db.whois.query.QueryMessages;
-import net.ripe.db.whois.query.VersionDateTime;
+import net.ripe.db.whois.common.dao.VersionDateTime;
 import net.ripe.db.whois.common.rpsl.ObjectType;
 
 import java.io.IOException;
